Comment and code cleaning
